Fixed damage animations 90% - NEED FIX
The animation will happen whenever a beast is attacked and will dissapear. However the only issue left, is that the animation only happens 1 out of 2 times. Everything that has to do with handling the animation is inside DamageOutput.cs

I'm using the variable isActive as a Condition to handle the animations, you can control the animation from the animator screen in unity.