Added a Audio System
Added Two New Scripts Sound.cs & AudioManager.cs

Sound.cs: Custom Script, this is made to hold a Sound Object for the AudioManager

---------------------------- This bad boy handles Sound that has to go from one scene to the other ----------------------------------------------------------

AudioManager.cs: This is what handles all the sound in the game, all the sounds are put into Sound[] Array, whenever adding a new sound, you need to use the inspector to add manually a new sound, and DONT FORGET to give it the same name as your target scene.

AudioManager -> Awake() --> The first part  (DontDestroyOnLoad) Makes sure the Music Gameobject doesn't get destroyer when switching scenes
                        -> Foreach, this just initialises the Sounds from the array.

AudioManager -> Play() --> This receives a String (NAME OF THE TARGETTED SCENE) , Looks for a Sound in the Sound Array with the same name as the Scene, Checks if null, then plays the sound.