Added theme songs for Main menus and Battle (ONLY MANOLI)
The theme songs are stored in an Audio Clip object, the audio source component of the Music game object gets updated on screen switch. Inside the Play() Method , there is an if/elseif/else statement that controls which clip gets played depending on the name of the scene.

The variable tempName is here to hold the previous scene, because when you need to switch from one song to another you need to switch audio clips, and replay.