Added Sound on Buttons For Summon page (SummonMain Scene)
Both buttons will Trigger an onclick , it calls the Audio source Component of the Object Button_Sound on the SummonMain scene.  Then calls the method PlayOneShot and plays the audio file linked to the Button call.

It's important to keep the Button_Sound object and it's audio source, the Audio source shoudn't be on the Button Object itself, otherwise it will just destroy itself when switching pages.